* better UI behavior in case stuff gets deleted * structural refactorings * warning icons for using deleted scope and ammo in rifle config

+++ Sharp.Ballistics.Calculator/ViewModels/AmmoViewModel.cs
@@ -6,12 +6,14 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 
 namespace Sharp.Ballistics.Calculator.ViewModels
 {
     public class AmmoViewModel : FunctionScreen
     {
-        private readonly IEventAggregator eventAggregator;
+        private readonly RiflesModel riflesModel;
+        private readonly IEventAggregator eventsAggregator;
 
         public override string IconFilename => "ammo.png";
 
@@ -21,10 +23,11 @@ namespace Sharp.Ballistics.Calculator.ViewModels
         private readonly IWindowManager windowManager;
         private readonly ConfigurationModel configurationModel;
         public AmmoViewModel(AmmoModel ammoModel, 
+                             RiflesModel riflesModel,
                              ConfigurationModel configurationModel, 
                              IWindowManager windowManager,
-                             IEventAggregator eventAggregator)
-            :base(eventAggregator)
+                             IEventAggregator eventsAggregator)
+            :base(eventsAggregator)
         {            
 #pragma warning disable CC0021 // Use nameof
             DisplayName = "Cartridges";
@@ -34,10 +37,11 @@ namespace Sharp.Ballistics.Calculator.ViewModels
             this.windowManager = windowManager;
 
             configurationModel.Initialize();
-            this.eventAggregator = eventAggregator;
+            this.eventsAggregator = eventsAggregator;
+            this.riflesModel = riflesModel;
         }
 
-        public UnitsConfiguration Units => configurationModel.Units;
+        public UnitSettings Units => configurationModel.Units;
 
         public IEnumerable<Cartridge> Cartridges => ammoModel.All();
 
@@ -47,7 +51,7 @@ namespace Sharp.Ballistics.Calculator.ViewModels
 
         public void AddCartridge()
         {
-            var newCartridgeViewModel = new EditCartridgeViewModel(configurationModel);
+            var newCartridgeViewModel = new EditCartridgeViewModel(configurationModel, ammoModel);
             if (windowManager.ShowDialog(newCartridgeViewModel) ?? false)
             {
                 Task.Run(() =>
@@ -69,7 +73,7 @@ namespace Sharp.Ballistics.Calculator.ViewModels
 
         public void EditCartridge(Cartridge cartridge)
         {
-            var editCartridgeViewModel = new EditCartridgeViewModel(configurationModel,cartridge);
+            var editCartridgeViewModel = new EditCartridgeViewModel(configurationModel,ammoModel,cartridge);
             if (windowManager.ShowDialog(editCartridgeViewModel) ?? false)
             {
                 Task.Run(() =>
@@ -91,8 +95,24 @@ namespace Sharp.Ballistics.Calculator.ViewModels
 
         public void RemoveCartridge(Cartridge cartridge)
         {
+            var relevantRifles = riflesModel.RiflesByCartridgeName(cartridge.Name).ToList();
+            if(relevantRifles.Any())
+            {
+                if (MessageBox.Show($"Deleting {cartridge.Name} will affect {relevantRifles.Count} stored rifle(s). Continue?","Query",
+                            MessageBoxButton.YesNo,
+                            MessageBoxImage.Warning) == MessageBoxResult.No)
+                    return;
+            }           
+
             ammoModel.Delete(cartridge);
             NotifyOfPropertyChange(() => Cartridges);
+            var appEvent = new AppEvent
+            {
+                Type = Constants.CartridgeRemovedMessage,
+            };
+            appEvent.Parameters.Add(Constants.ChangedItemName, cartridge.Name);
+
+            Messenger.PublishOnBackgroundThread(appEvent);
         }
