+++ Sharp.Ballistics.Calculator/ViewModels/ScopesViewModel.cs
@@ -2,12 +2,15 @@
 using Sharp.Ballistics.Abstractions;
 using Sharp.Ballistics.Calculator.Models;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
+using System.Windows;
 
 namespace Sharp.Ballistics.Calculator.ViewModels
 {
     public class ScopesViewModel : FunctionScreen
     {
+        private readonly RiflesModel riflesModel;
         private readonly ConfigurationModel configurationModel;
         private readonly IWindowManager windowManager;
         private readonly ScopesModel scopesModel;
@@ -21,20 +24,22 @@ namespace Sharp.Ballistics.Calculator.ViewModels
 
         public ScopesViewModel(ScopesModel scopesModel,
             ConfigurationModel configurationModel,
+            RiflesModel riflesModel,
             IWindowManager windowManager,
-            IEventAggregator eventAggregator) : base(eventAggregator)
+            IEventAggregator eventsAggregator) : base(eventsAggregator)
         {
             this.scopesModel = scopesModel;
             this.windowManager = windowManager;
             this.configurationModel = configurationModel;
             DisplayName = "Scopes";
+            this.riflesModel = riflesModel;
         }
 
         public IEnumerable<Scope> Scopes => scopesModel.All();
 
         public void AddScope()
         {
-            var newScopeViewModel = new EditScopeViewModel(configurationModel);
+            var newScopeViewModel = new EditScopeViewModel(configurationModel,scopesModel);
             if (windowManager.ShowDialog(newScopeViewModel) ?? false)
             {
                 Task.Run(() =>
@@ -56,7 +61,7 @@ namespace Sharp.Ballistics.Calculator.ViewModels
 
         public void EditScope(Scope scope)
         {
-            var editScopeViewModel = new EditScopeViewModel(configurationModel, scope);
+            var editScopeViewModel = new EditScopeViewModel(configurationModel,scopesModel, scope);
             if (windowManager.ShowDialog(editScopeViewModel) ?? false)
             {
                 Task.Run(() =>
@@ -78,8 +83,25 @@ namespace Sharp.Ballistics.Calculator.ViewModels
 
         public void RemoveScope(Scope scope)
         {
+            var relevantRifles = riflesModel.RiflesByScopeName(scope.Name).ToList();
+            if (relevantRifles.Any())
+            {
+                if (MessageBox.Show($"Deleting {scope.Name} will affect {relevantRifles.Count} stored rifle(s). Continue?", "Query",
+                            MessageBoxButton.YesNo,
+                            MessageBoxImage.Warning) == MessageBoxResult.No)
+                    return;
+            }
+
             scopesModel.Delete(scope);
             NotifyOfPropertyChange(() => Scopes);
+
+            var appEvent = new AppEvent
+            {
+                Type = Constants.ScopeRemovedMessage,
+            };
+            appEvent.Parameters.Add(Constants.ChangedItemName, scope.Name);
+
+            Messenger.PublishOnBackgroundThread(appEvent);
         }
     }
 }
